relocate runtime modules to scripts folder
updated require statements, will test
