+++ control.lua
@@ -1,32 +1,32 @@
 --Main file for mod runtime
 local util = require('util')
-local fa_utils = require('fa-utils')
-local fa_localising = require('localising')
-local fa_crafting = require("crafting")
-local fa_electrical = require("electrical")
-local fa_equipment = require("equipment")
-local fa_combat = require("combat")
-local fa_graphics = require("graphics")
-local fa_mouse = require("mouse")
-local fa_tutorial = require("tutorial-system")
-local fa_sectors = require("building-vehicle-sectors")
-local fa_menu_search = require("menu-search")
-local fa_building_tools = require("building-tools")
-local fa_mining_tools = require("mining-tools")
-local fa_rails = require("rails")
-local fa_rail_builder = require("rail-builder")
-local fa_trains = require("trains")
-local fa_train_stops = require("train-stops")
-local fa_driving = require("driving")
-local fa_scanner = require("scanner")
-local fa_spidertrons = require("spidertron")
-local fa_belts = require("transport-belts")
-local fa_zoom = require('zoom')
-local fa_bot_logistics = require("worker-robots")
-local fa_blueprints = require("blueprints")
-local fa_travel = require("travel-tools")
-local fa_teleport = require("teleport")
-local fa_warnings = require("warnings")
+local fa_utils = require('scripts.fa-utils')
+local fa_localising = require('scripts.localising')
+local fa_crafting = require("scripts.crafting")
+local fa_electrical = require("scripts.electrical")
+local fa_equipment = require("scripts.equipment")
+local fa_combat = require("scripts.combat")
+local fa_graphics = require("scripts.graphics")
+local fa_mouse = require("scripts.mouse")
+local fa_tutorial = require("scripts.tutorial-system")
+local fa_sectors = require("scripts.building-vehicle-sectors")
+local fa_menu_search = require("scripts.menu-search")
+local fa_building_tools = require("scripts.building-tools")
+local fa_mining_tools = require("scripts.mining-tools")
+local fa_rails = require("scripts.rails")
+local fa_rail_builder = require("scripts.rail-builder")
+local fa_trains = require("scripts.trains")
+local fa_train_stops = require("scripts.train-stops")
+local fa_driving = require("scripts.driving")
+local fa_scanner = require("scripts.scanner")
+local fa_spidertrons = require("scripts.spidertron")
+local fa_belts = require("scripts.transport-belts")
+local fa_zoom = require('scripts.zoom')
+local fa_bot_logistics = require("scripts.worker-robots")
+local fa_blueprints = require("scripts.blueprints")
+local fa_travel = require("scripts.travel-tools")
+local fa_teleport = require("scripts.teleport")
+local fa_warnings = require("scripts.warnings")
 
 local circuit_networks = require('circuit-networks')
