+++ scripts/electrical.lua
@@ -1,6 +1,6 @@
 --Here: Electricity related functions and menus
 local util = require("util")
-local fa_utils = require("fa-utils")
+local fa_utils = require("scripts.fa-utils")
 
 local mod = {}
