+++ scripts/menu-search.lua
@@ -1,9 +1,9 @@
 --Here: Menu search and directly related functions
 local util = require('util')
-local fa_utils = require('fa-utils')
-local fa_crafting = require("crafting")
-local localising = require('localising')
-local fa_sectors = require("building-vehicle-sectors")
+local fa_utils = require('scripts.fa-utils')
+local fa_crafting = require("scripts.crafting")
+local localising = require('scripts.localising')
+local fa_sectors = require("scripts.building-vehicle-sectors")
 
 local mod = {}
