+++ scripts/mining-tools.lua
