+++ scripts/rail-builder.lua
@@ -3,9 +3,9 @@
 ---@diagnostic disable: assign-type-mismatch
 
 local util = require('util')
-local fa_utils = require('fa-utils')
-local fa_mining_tools = require("mining-tools")
-local fa_rails = require("rails")
+local fa_utils = require("scripts.fa-utils")
+local fa_mining_tools = require("scripts.mining-tools")
+local fa_rails = require("scripts.rails")
 local dirs = defines.direction
 
 local mod = {}
