+++ scripts/rails.lua
@@ -2,7 +2,7 @@
 --Does not include event handlers
 
 local util = require('util')
-local fa_utils = require('fa-utils')
+local fa_utils = require('scripts.fa-utils')
 local dirs = defines.direction
 
 local mod = {}
