+++ scripts/spidertron.lua
