+++ scripts/teleport.lua
@@ -1,7 +1,7 @@
 --Here: teleporting
-local fa_utils = require("fa-utils")
-local fa_graphics = require("graphics")
-local fa_mouse = require("mouse")
+local fa_utils = require("scripts.fa-utils")
+local fa_graphics = require("scripts.graphics")
+local fa_mouse = require("scripts.mouse")
 
 local mod = {}
