+++ scripts/transport-belts.lua
@@ -1,8 +1,8 @@
 --Here: functions about belts, splitters, underground belts
 
-local localising = require("localising")
+local localising = require("scripts.localising")
 local util = require('util')
-local fa_utils = require('fa-utils')
+local fa_utils = require('scripts.fa-utils')
 
 local mod = {}
